requirements.txt
numpy
pandas
matplotlib
scikit-learn
streamlit
joblib
plotly
reportlab
streamlit-option-menu
flask
flask-cors
pymongo
google-generativeai
groq
sendgrid==6.11.0
python-dotenv==1.0.0
psutil
